service layer tested >80 percent
